fix: Historychart error after refactor
